Initial naive impl of inventory endpoints
* Modified data_gatherer so that last_updated is in a separate object in
each collection, and is only updated at the very end if primes or relics
were inserted/updated. This is to match the data model of the inventory,
which also only has last_updated as a separate object (instead of as an
element of every object in the collection).

* Added helper functions to get and set said last_updated object in
Flask. Reworked the universal_data endpoint to use those functions.

* Also added missing relics list to the universal_data endpoint. Oops.

* Added inventory endpoints (GET/PUT). For now, the PUT is naive and
just dumps whatever the user sends into the DB. We should probably
sanitize it later.