T85+102+105: The big one; Google Auth implemented!
T85: Google Auth implemented for dev and prod environments. It currently
works about as well as I want it to, though I still need to style the
button a bit.

T102: Moved Flask's secret and a bunch of hard-coded URIs into separate
files. Flask's secret uses docker-compose's env_file functionality.

T105: Allow simple auth in dev.

Note that 85 was the main focus, but 102 and 105 needed to be completed
as a result.

https://trac.ibbathon.com/trac/ticket/85
https://trac.ibbathon.com/trac/ticket/102
https://trac.ibbathon.com/trac/ticket/105