T139: Switched to official drop tables
Now all my prime/relic data is coming from the official drop tables.
Next up is providing a way to update the needed/ducat values for prime
parts, as the official drop tables don't provide that. It's rare
enough they differ from the defaults, so I don't want to set up a whole
other scraper for them.

Also fixed the warning about new URL parser and a few minor tweaks to
style, as well as putting in the groundwork for better process reporting.

https://trac.ibbathon.com/trac/ticket/139