T147: Fixed issue with users overwriting others data
I hadn't updated the queries for the PUTs to include user_id after I
switched to a multi-user environment. That was a long time ago...

Also switched to using email as the auth ID, as it will make debugging
easier.

Also added a way to invalidate sessions on the Mongo side in case that's
ever desired.

https://trac.ibbathon.com/trac/ticket/147